ACF RBAC 完成

+++ advanced/backend/controllers/PostController.php
@@ -7,6 +7,7 @@ use common\models\Post;
 use common\models\PostSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
 
 /**
@@ -24,7 +25,25 @@ class PostController extends Controller {
                 'actions' => [
                     'delete' => ['POST'],
                 ],
+                ],
+            'access'=>[
+                'class'=> \yii\filters\AccessControl::className(),
+                'rules'=>
+                [
+                    [
+                        'actions'=>['index','view'],
+                        'allow'=>true,
+                        'roles'=>['?'],
+                    ],
+                    [
+                        'actions'=>['view','index','create','update'],
+                        'allow'=>true,
+                        'roles'=>['@'],
+                    ]
+                ],
+                
             ],
+      
         ];
     }
 
@@ -59,6 +78,11 @@ class PostController extends Controller {
      * @return mixed
      */
     public function actionCreate() {
+        
+        if(!Yii::$app->user->can('createPost')){
+            throw new ForbiddenHttpException('对不起，你没有进行该操作的权限');
+        }
+        
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -77,6 +101,11 @@ class PostController extends Controller {
      * @return mixed
      */
     public function actionUpdate($id) {
+        
+        if(!Yii::$app->user->can('updatePost')){
+            throw new ForbiddenHttpException('对不起，你没有进行该操作的权限');
+        }
+        
         $model = $this->findModel($id);
         
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -95,6 +124,11 @@ class PostController extends Controller {
      * @return mixed
      */
     public function actionDelete($id) {
+        
+        if(!Yii::$app->user->can('deletePost')){
+            throw new ForbiddenHttpException('对不起，你没有进行该操作的权限');
+        }        
+        
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
